+++ yash-builtin/src/common/arg.rs
@@ -326,11 +326,11 @@ impl<'a> From<&'a Error<'_>> for Message<'a> {
     fn from(error: &'a Error<'_>) -> Self {
         let field = error.field();
 
-        let mut a = vec![Annotation {
-            r#type: AnnotationType::Error,
-            label: field.value.as_str().into(),
-            location: &field.origin,
-        }];
+        let mut a = vec![Annotation::new(
+            AnnotationType::Error,
+            field.value.as_str().into(),
+            &field.origin,
+        )];
 
         field.origin.code.source.complement_annotations(&mut a);