+++ yash-syntax/src/parser/error.rs
@@ -424,27 +424,27 @@ impl fmt::Display for Error {
 
 impl<'a> From<&'a Error> for Message<'a> {
     fn from(e: &'a Error) -> Self {
-        let mut a = vec![Annotation {
-            r#type: AnnotationType::Error,
-            label: e.cause.label().into(),
-            location: &e.location,
-        }];
+        let mut a = vec![Annotation::new(
+            AnnotationType::Error,
+            e.cause.label().into(),
+            &e.location,
+        )];
 
         e.location.code.source.complement_annotations(&mut a);
 
         if let Some((location, label)) = e.cause.related_location() {
-            a.push(Annotation {
-                r#type: AnnotationType::Info,
-                label: label.into(),
+            a.push(Annotation::new(
+                AnnotationType::Info,
+                label.into(),
                 location,
-            });
+            ));
         }
         if let ErrorCause::Syntax(SyntaxError::BangAfterBar) = &e.cause {
-            a.push(Annotation {
-                r#type: AnnotationType::Help,
-                label: "surround this in a grouping: `{ ! ...; }`".into(),
-                location: &e.location,
-            })
+            a.push(Annotation::new(
+                AnnotationType::Help,
+                "surround this in a grouping: `{ ! ...; }`".into(),
+                &e.location,
+            ));
         }
 
         Message {